refactor cors, address pr comments
