+++ onboarding-api/src/controllers/kyb/kybCallback.ts
@@ -1,4 +1,3 @@
-import * as crypto from 'crypto'
 import { Request, Response } from 'express'
 import { InferType, mixed, object, string } from 'yup'
 import { onboardingBucket, OnboardingUser, validateAndWriteToFirestore } from '../../database'
@@ -7,37 +6,8 @@ import { sendVerifiedBusinessMessage } from '../../emails/sendVerifiedBusinessMe
 import { fetchUser } from '../../utils/fetchUser'
 import { HttpError, reportHttpError } from '../../utils/httpError'
 import { shuftiProRequest } from '../../utils/shuftiProRequest'
-import { Subset, SupportedNetworks } from '../../utils/types'
+import { KybCallbackRequestBody, Subset, SupportedNetworks } from '../../utils/types'
 import { validateInput } from '../../utils/validateInput'
-
-type VerificationState = 1 | 0 | null
-
-type RequestBody = {
-  reference: `KYB_${string}`
-  event:
-    | `request.${'pending' | 'timeout' | 'deleted' | 'received'}`
-    | 'review.pending'
-    | `verification.${'accepted' | 'declined' | 'cancelled' | 'status.changed'}`
-  verification_url: `https://app.shuftipro.com/verification/process/${string}`
-  email: string
-  country: string
-
-  /**
-   * This object will be returned in case of verification.accepted or verification.declined.
-   * This object will include all the gathered data in a request process.
-   */
-  verification_data?: unknown
-  verification_result?: {
-    proof_stores: {
-      articles_of_association: VerificationState
-      certificate_of_incorporation: VerificationState
-      proof_of_address: VerificationState
-      register_of_directors: VerificationState
-      register_of_shareholders: VerificationState
-      signed_and_dated_ownership_structure: VerificationState
-    }
-  }
-}
 
 const kybCallbackInput = object({
   address: string().required(),
@@ -47,17 +17,11 @@ const kybCallbackInput = object({
 })
 
 export const kybCallbackController = async (
-  req: Request<any, any, RequestBody, InferType<typeof kybCallbackInput>>,
+  req: Request<any, any, KybCallbackRequestBody, InferType<typeof kybCallbackInput>>,
   res: Response
 ) => {
   try {
-    const { headers, body, query } = req
-
-    const isValidRequest = headers.signature && isValidShuftiRequest(body, headers.signature)
-
-    if (!isValidRequest) {
-      throw new HttpError(401, 'Unauthorized')
-    }
+    const { body, query } = req
 
     await validateInput(query, kybCallbackInput)
 
@@ -81,7 +45,7 @@ export const kybCallbackController = async (
       return res.status(200).end()
     }
 
-    const status: RequestBody = await shuftiProRequest({ reference: body.reference }, { path: 'status', dryRun: false })
+    const status = await shuftiProRequest({ reference: body.reference }, { path: 'status', dryRun: false })
 
     if (status.event === 'verification.declined') {
       await sendVerifiedBusinessMessage(user.email, false, query.poolId, query.trancheId)
@@ -123,24 +87,6 @@ export const kybCallbackController = async (
     const error = reportHttpError(e)
     return res.status(error.code).send({ error: error.message })
   }
-}
-
-function isValidShuftiRequest(body: RequestBody, signature: string | string[]) {
-  const requestBody = JSON.stringify(body)
-    //  escape all `/`
-    .replace(/\//g, '\\/')
-
-    // replace greek characters with unicodes
-    .replace(/\u00f4/g, '\\u00f4')
-    .replace(/\u00fa/g, '\\u00fa')
-    .replace(/\u039a/g, '\\u039a')
-    .replace(/\u039d/g, '\\u039d')
-    .replace(/\u03a4/g, '\\u03a4')
-    .replace(/\u0399/g, '\\u0399')
-
-  const hash = crypto.createHash('sha256').update(`${requestBody}${process.env.SHUFTI_PRO_SECRET_KEY}`).digest('hex')
-
-  return hash === signature
 }
 
 async function fetchSignedAgreement(wallet: Request['wallet'], poolId: string, trancheId: string) {
