+++ onboarding-api/src/middleware/cors.ts
@@ -1,18 +1,19 @@
 const cors = require('cors')
 
-const centrifugeDomains = [
+const allowedOrigins = [
   /^(https:\/\/.*cntrfg\.com)$/,
   /^(https:\/\/.*centrifuge\.io)$/,
   /^(https:\/\/.*altair\.network)$/,
   /^(https:\/\/.*k-f\.dev)$/,
+  /^(https:\/\/shuftipro\.com)$/,
+  /^(http:\/\/localhost:)\d{1,5}$/,
 ]
 
 export const corsMiddleware = cors({
   origin: (origin, callback) => {
-    const isLocalhost = /^(http:\/\/localhost:)\d{1,5}$/.test(origin)
-    const isCentrifugeDomain = centrifugeDomains.some((regex) => regex.test(origin))
+    const isAllowedOrigin = allowedOrigins.some((regex) => regex.test(origin))
 
-    if (isLocalhost || isCentrifugeDomain || origin === undefined) {
+    if (isAllowedOrigin) {
       callback(null, true)
     } else {
       callback(new Error(`Not allowed by CORS, ${origin}`))
