+++ onboarding-api/src/utils/types.d.ts
@@ -36,3 +36,32 @@ declare global {
     }
   }
 }
+
+type VerificationState = 1 | 0 | null
+
+export type KybCallbackRequestBody = {
+  reference: `KYB_${string}`
+  event:
+    | `request.${'pending' | 'timeout' | 'deleted' | 'received'}`
+    | 'review.pending'
+    | `verification.${'accepted' | 'declined' | 'cancelled' | 'status.changed'}`
+  verification_url: `https://app.shuftipro.com/verification/process/${string}`
+  email: string
+  country: string
+
+  /**
+   * This object will be returned in case of verification.accepted or verification.declined.
+   * This object will include all the gathered data in a request process.
+   */
+  verification_data?: unknown
+  verification_result?: {
+    proof_stores: {
+      articles_of_association: VerificationState
+      certificate_of_incorporation: VerificationState
+      proof_of_address: VerificationState
+      register_of_directors: VerificationState
+      register_of_shareholders: VerificationState
+      signed_and_dated_ownership_structure: VerificationState
+    }
+  }
+}
